requirements.txt
aioredis==1.2.0
aniso8601==3.0.2
asgiref==2.3.2
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
hiredis==0.3.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.6.0
oauthlib==3.0.0
Pillow==5.4.1
promise==2.2.1
PyHamcrest==1.9.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.0
zope.interface==4.6.0
